Decoupling the handlers from the index file
Defining the handlers in the index file sucks because for a few reasons. If we
want to change the handler associated with a path/route, we're going to have
edit the code directly. Secondly, perhaps we would like to reuse that handler
elsewhere, perhaps in a different index, we would have to rewrite it. That's not
very DRY. Thirdly, these handlers might have dependencies which in turn makes
the entire index file dependent on these dependencies. Perhaps, we'd like to
make our index as general as possible so that it can be re-used independently
(as module). In this case, we would use injection dependencies to pass in an
object as an argument instead of importing a module. However, this index is
specific to our application, so we don't have to make it as general as possible.

As a solution to all of these problems, we decouple the handlers from the index
file, and we place the handlers in a standalone module that can be imported for
use. This way if we need to change all the handlers, we can just import a
different index file. Hell, we can even have multiple handler modules imported,
each with a specific purpose. We are also free to reuse these handlers by simply
referencing them.